Rename wagon project for Artifact Registry.
Rename packages, directories, modules, classes from Cloud Build
Artifacts to Artifact Registry. Update documentation to refer to
Artifact Registry.

Version is bumped to 2.0.0-SNAPSHOT.

Adds .gitignore file.
Updates the gradle.build so that signing only happens on releases.

Rename buildartifacts-auth-common module and package to artifactregistry-auth-common.

Add gitignore

Rename buildartifacts-gradle-plugin to artifactregistry-gradle-plugin

Rename buildartifacts-maven-wagon to artifactregistry-maven-wagon

Update version to 2.0.0-SNAPSHOT. Substitute all remaining references to buildartifacts to be artifactregistry, including docs.

Configure gradle so that it doesn't do signing for SNAPSHOT releases
