+++ artifactregistry-gradle-plugin/src/main/resources/META-INF/gradle-plugins/com.google.cloud.artifactregistry.gradle-credential-helper.properties
@@ -1,0 +1,1 @@
+implementation-class=com.google.cloud.artifactregistry.gradle.plugin.ArtifactRegistryGradlePlugin
