+++ artifactregistry-maven-wagon/src/main/resources/META-INF/plexus/components.xml
@@ -1,0 +1,10 @@
+<component-set>
+  <components>
+    <component>
+      <role>org.apache.maven.wagon.Wagon</role>
+      <role-hint>artifactregistry</role-hint>
+      <implementation>com.google.cloud.artifactregistry.wagon.ArtifactRegistryWagon</implementation>
+      <instantiation-strategy>per-lookup</instantiation-strategy>
+    </component>
+  </components>
+</component-set>
