--- buildartifacts-gradle-plugin/build.gradle
@@ -1,26 +1,0 @@
-plugins {
-    id 'java-gradle-plugin' 
-    id 'com.gradle.plugin-publish' version '0.10.0' 
-}
-
-dependencies {
-	compile gradleApi()
-	compile group: 'com.google.auth', name: 'google-auth-library-oauth2-http', version: '0.16.2'
-}
-
-pluginBundle {
-    website = 'https://github.com/GoogleCloudPlatform/build-artifacts-maven-tools' 
-    vcsUrl = 'https://github.com/GoogleCloudPlatform/build-artifacts-maven-tools' 
-    tags = ['buildartifacts', 'gcp'] 
-}
-
-gradlePlugin {
-    plugins { 
-        cloudBuildArtifactsPlugin { 
-            id = 'com.google.cloud.buildartifacts.gradle-plugin' 
-            displayName = 'Cloud Build Artifacts Gradle Plugin' 
-            description = 'A Gradle plugin used to connect to Cloud Build Artifacts Maven repositories.' 
-            implementationClass = 'com.google.cloud.buildartifacts.gradle.plugin.BuildArtifactsGradlePlugin'
-        }
-    }
-}
