--- buildartifacts-gradle-plugin/src/main/resources/META-INF/gradle-plugins/com.google.cloud.buildartifacts.gradle-credential-helper.properties
@@ -1,1 +1,0 @@
-implementation-class=com.google.cloud.buildartifacts.gradle.plugin.BuildArtifactsGradlePlugin
