--- buildartifacts-maven-wagon/src/main/resources/META-INF/plexus/components.xml
@@ -1,10 +1,0 @@
-<component-set>
-  <components>
-    <component>
-      <role>org.apache.maven.wagon.Wagon</role>
-      <role-hint>buildartifacts</role-hint>
-      <implementation>com.google.cloud.buildartifacts.wagon.BuildArtifactsWagon</implementation>
-      <instantiation-strategy>per-lookup</instantiation-strategy>
-    </component>
-  </components>
-</component-set>
